+++ Panels/Panel.js
@@ -6,6 +6,7 @@ import Slottable from '@enact/ui/Slottable';
 import PropTypes from 'prop-types';
 import React from 'react';
 
+import Skinnable from '../Skinnable';
 import SharedStateDecorator from '../internal/SharedStateDecorator';
 
 import {PanelTypeContext} from './Viewport';
@@ -208,7 +209,9 @@ const Panel = SharedStateDecorator(
 		},
 		Slottable(
 			{slots: ['header']},
-			PanelBase
+			Skinnable(
+				PanelBase
+			)
 		)
 	)
 );